+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/Account.kt
@@ -34,7 +34,6 @@ import com.vitorpamplona.amethyst.commons.connectedApps.signers.NostrSignerPermi
 import com.vitorpamplona.amethyst.commons.marmot.MarmotManager
 import com.vitorpamplona.amethyst.commons.model.IAccount
 import com.vitorpamplona.amethyst.commons.model.buzz.BuzzRelayDialect
-import com.vitorpamplona.amethyst.commons.model.buzz.WorkflowRunPayload
 import com.vitorpamplona.amethyst.commons.model.concord.ConcordChannel
 import com.vitorpamplona.amethyst.commons.model.concord.ConcordChannelListState
 import com.vitorpamplona.amethyst.commons.model.concord.ConcordSessionManager
@@ -47,10 +46,8 @@ import com.vitorpamplona.amethyst.commons.model.nip28PublicChats.PublicChatChann
 import com.vitorpamplona.amethyst.commons.model.nip28PublicChats.PublicChatListDecryptionCache
 import com.vitorpamplona.amethyst.commons.model.nip28PublicChats.PublicChatListState
 import com.vitorpamplona.amethyst.commons.model.nip29RelayGroups.RelayGroupChannel
-import com.vitorpamplona.amethyst.commons.model.nip29RelayGroups.RelayGroupDeletions
 import com.vitorpamplona.amethyst.commons.model.nip29RelayGroups.RelayGroupListDecryptionCache
 import com.vitorpamplona.amethyst.commons.model.nip29RelayGroups.RelayGroupListState
-import com.vitorpamplona.amethyst.commons.model.nip29RelayGroups.RelayGroupMembership
 import com.vitorpamplona.amethyst.commons.model.nip30CustomEmojis.EmojiPackState
 import com.vitorpamplona.amethyst.commons.model.nip38UserStatuses.UserStatusAction
 import com.vitorpamplona.amethyst.commons.model.nip51Lists.favoriteAlgoFeedsLists.FavoriteAlgoFeedsListDecryptionCache
@@ -160,26 +157,9 @@ import com.vitorpamplona.amethyst.service.uploads.FileHeader
 import com.vitorpamplona.amethyst.ui.actions.NewMessageTagger
 import com.vitorpamplona.amethyst.ui.navigation.bottombars.BottomBarEntry
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.EventProcessor
-import com.vitorpamplona.quartz.buzz.dm.DmAddMemberEvent
-import com.vitorpamplona.quartz.buzz.dm.DmHideEvent
-import com.vitorpamplona.quartz.buzz.dm.DmOpenEvent
-import com.vitorpamplona.quartz.buzz.jobs.JobCancelEvent
-import com.vitorpamplona.quartz.buzz.jobs.JobRequestEvent
-import com.vitorpamplona.quartz.buzz.presence.TypingIndicatorEvent
-import com.vitorpamplona.quartz.buzz.relayAdmin.RelayAdminAddMemberEvent
-import com.vitorpamplona.quartz.buzz.relayAdmin.RelayAdminRemoveMemberEvent
 import com.vitorpamplona.quartz.buzz.threading.buzzThread
 import com.vitorpamplona.quartz.buzz.threading.buzzThreadReply
 import com.vitorpamplona.quartz.buzz.threading.buzzThreadRoot
-import com.vitorpamplona.quartz.buzz.workflow.ApprovalDenyEvent
-import com.vitorpamplona.quartz.buzz.workflow.ApprovalGrantEvent
-import com.vitorpamplona.quartz.buzz.workflow.WorkflowDefEvent
-import com.vitorpamplona.quartz.buzz.workflow.WorkflowTriggerEvent
-import com.vitorpamplona.quartz.buzz.workflow.workflowChannel
-import com.vitorpamplona.quartz.buzz.workspace.BUZZ_ROLE_ADMIN
-import com.vitorpamplona.quartz.buzz.workspace.BUZZ_ROLE_MEMBER
-import com.vitorpamplona.quartz.buzz.workspace.BUZZ_VISIBILITY_OPEN
-import com.vitorpamplona.quartz.buzz.workspace.BUZZ_VISIBILITY_PRIVATE
 import com.vitorpamplona.quartz.concord.cord03Channels.ChannelChat
 import com.vitorpamplona.quartz.concord.cord03Channels.ConcordChannelId
 import com.vitorpamplona.quartz.experimental.bounties.BountyAddValueEvent
@@ -214,10 +194,7 @@ import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.crypto.KeyPair
 import com.vitorpamplona.quartz.nip01Core.hints.EventHintBundle
 import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.PublishResult
-import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.fetchAllWithHooks
 import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.fetchFirst
-import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.publishAndCollectResults
 import com.vitorpamplona.quartz.nip01Core.relay.client.paging.RelayLoadingCursors
 import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
@@ -268,20 +245,8 @@ import com.vitorpamplona.quartz.nip19Bech32.entities.NSec
 import com.vitorpamplona.quartz.nip22Comments.CommentEvent
 import com.vitorpamplona.quartz.nip22Comments.notify
 import com.vitorpamplona.quartz.nip25Reactions.ReactionEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.GroupId
 import com.vitorpamplona.quartz.nip29RelayGroups.hTag
-import com.vitorpamplona.quartz.nip29RelayGroups.metadata.GroupMetadataEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.CreateGroupEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.CreateInviteEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.DeleteGroupEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.EditMetadataEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.PutUserEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.RemoveUserEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.moderation.UpdatePinListEvent
 import com.vitorpamplona.quartz.nip29RelayGroups.moderation.previous
-import com.vitorpamplona.quartz.nip29RelayGroups.request.JoinRequestEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.request.LeaveRequestEvent
-import com.vitorpamplona.quartz.nip29RelayGroups.tags.GroupIdTag
 import com.vitorpamplona.quartz.nip32Labeling.LabelEvent
 import com.vitorpamplona.quartz.nip36SensitiveContent.contentWarning
 import com.vitorpamplona.quartz.nip37Drafts.DraftEventCache
@@ -332,7 +297,6 @@ import com.vitorpamplona.quartz.nip72ModCommunities.rules.CommunityRulesEvent
 import com.vitorpamplona.quartz.nip72ModCommunities.rules.tags.KindRuleTag
 import com.vitorpamplona.quartz.nip72ModCommunities.rules.tags.PubkeyRuleTag
 import com.vitorpamplona.quartz.nip72ModCommunities.rules.tags.WotTag
-import com.vitorpamplona.quartz.nip7DThreads.ThreadEvent
 import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
 import com.vitorpamplona.quartz.nip88Polls.response.PollResponseEvent
 import com.vitorpamplona.quartz.nip89AppHandlers.clientTag.NostrSignerWithClientTag
@@ -378,8 +342,6 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import java.math.BigDecimal
 import kotlin.coroutines.cancellation.CancellationException
 import com.vitorpamplona.quartz.experimental.nip95.header.thumbhash as nip95thumbhash
@@ -760,6 +722,9 @@ class Account(
 
     /** Marmot/MLS group orchestration (create/members/admins/messages/key packages). */
     val marmot = AccountMarmotActions(this)
+
+    /** NIP-29 relay-group + Buzz workspace orchestration. */
+    val relayGroups = AccountRelayGroupActions(this)
 
     /**
      * Relay routing + sign-and-publish choke point: computes which relays an event
@@ -2079,471 +2044,6 @@ class Account(
         if (imetas.isEmpty()) return text
         val extraUrls = imetas.map { it.url }.filter { it.isNotBlank() && !text.contains(it) }
         return (listOf(text) + extraUrls).filter { it.isNotBlank() }.joinToString("\n")
-    }
-
-    suspend fun joinRelayGroup(
-        channel: RelayGroupChannel,
-        code: String? = null,
-    ) {
-        val template = JoinRequestEvent.build(channel.groupId.id, inviteCode = code)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-        follow(channel)
-    }
-
-    /**
-     * Fire a Buzz kind-20002 typing heartbeat for [channel] to its host relay. Ephemeral
-     * (never stored) and fire-and-forget — no delivery tracking, no local echo (we filter
-     * our own typing in the UI). Throttled by the composer to [BuzzTypingState.TYPING_HEARTBEAT_SECS].
-     */
-    suspend fun sendBuzzTyping(channel: RelayGroupChannel) {
-        if (!isWriteable()) return
-        val signed = signer.sign(TypingIndicatorEvent.build(channel.groupId.id))
-        client.publish(signed, setOf(channel.groupId.relayUrl))
-    }
-
-    /**
-     * Open (or re-surface) a Buzz DM with [participants] on [relay] via a kind-41010
-     * command. [participants] are the OTHER 1-8 people — the relay adds me, derives the
-     * canonical channel UUID, and confirms with a relay-signed [DmCreatedEvent]
-     * (kind-41001) that lands in [com.vitorpamplona.amethyst.commons.model.buzz.BuzzDmRegistry].
-     * We never assign the channel id ourselves, so callers discover the materialized DM
-     * by watching that registry rather than from this call's return.
-     */
-    suspend fun openBuzzDm(
-        relay: NormalizedRelayUrl,
-        participants: List<HexKey>,
-    ): String? {
-        val signed = signer.sign(DmOpenEvent.build(participants))
-        // The relay confirms the DM synchronously in the OK as `response:{"channel_id":"…"}` —
-        // the authoritative, relay-assigned channel UUID (the deployed relay does not emit a
-        // queryable kind-41001). Read it straight from the ack so the caller can open the chat.
-        var results = client.publishAndCollectResults(signed, setOf(relay))
-        var channelId = buzzDmChannelIdFromAck(results)
-
-        // NIP-42 write race: on a cold connection the relay rejects the first publish with
-        // `auth-required` (our AUTH reply lands async and the write path doesn't re-send). Warm
-        // the connection with a pendingOnAuthRequired read so the auth coordinator completes the
-        // handshake, then retry the publish on the now-authed socket. Mirrors the amy CLI fix.
-        if (channelId == null && results.values.any { !it.accepted && it.message.contains("auth-required", ignoreCase = true) }) {
-            client.fetchAllWithHooks(
-                filters = mapOf(relay to listOf(Filter(kinds = listOf(DmOpenEvent.KIND), limit = 1))),
-                timeoutMs = 8_000,
-                pendingOnAuthRequired = true,
-            ) { _, _ -> false }
-            results = client.publishAndCollectResults(signed, setOf(relay))
-            channelId = buzzDmChannelIdFromAck(results)
-        }
-        return channelId
-    }
-
-    /** The relay-assigned DM channel id from a DM-open OK message (`response:{"channel_id":"…"}`). */
-    private fun buzzDmChannelIdFromAck(results: Map<NormalizedRelayUrl, PublishResult>): String? =
-        results.values
-            .firstOrNull { it.accepted }
-            ?.message
-            ?.substringAfter("\"channel_id\":\"", "")
-            ?.substringBefore('"')
-            ?.takeIf { it.isNotBlank() }
-
-    /** Hide a Buzz DM from my sidebar with a kind-41012 command (re-opening it un-hides). */
-    suspend fun hideBuzzDm(channel: RelayGroupChannel) {
-        val template = DmHideEvent.build(channel.groupId.id)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /** Add [member] to an existing group DM with a kind-41011 command (creates a new DM set). */
-    suspend fun addBuzzDmMember(
-        channel: RelayGroupChannel,
-        member: HexKey,
-    ) {
-        val template = DmAddMemberEvent.build(channel.groupId.id, member)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /**
-     * File a Buzz agent job (kind-43001) into channel [channelId] on [relay] — a shared
-     * feature-request the workspace bot can pick up. Untargeted: any agent watching the
-     * channel may accept it. Returns the new job id (the request event id), or null when the
-     * account can't write. See [com.vitorpamplona.amethyst.commons.model.buzz.BuzzJobAggregator].
-     */
-    suspend fun fileBuzzJob(
-        relay: NormalizedRelayUrl,
-        channelId: String,
-        request: String,
-    ): HexKey? {
-        if (!isWriteable()) return null
-        val signed = signer.sign(JobRequestEvent.build(request, channelId, null))
-        // Reflect it locally so the board updates immediately (publish only sends to relays).
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-        return signed.id
-    }
-
-    /** Cancel a Buzz job [jobId] with a kind-43005 scoped to [channelId] on [relay]. */
-    suspend fun cancelBuzzJob(
-        relay: NormalizedRelayUrl,
-        channelId: String,
-        jobId: HexKey,
-    ) {
-        if (!isWriteable()) return
-        val signed = signer.sign(JobCancelEvent.build(jobId, "", channelId))
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-    }
-
-    /**
-     * Trigger a Buzz **workflow** run (kind-46020) for [workflowId] into channel [channelId] on
-     * [relay], carrying [task] as the run's request. The trigger's event id IS the run id (and the
-     * approval token), returned here. A run pauses on a human-approval gate before anything ships —
-     * see [com.vitorpamplona.amethyst.commons.model.buzz.WorkflowRunAggregator].
-     */
-    suspend fun triggerBuzzWorkflow(
-        relay: NormalizedRelayUrl,
-        channelId: String,
-        workflowId: String,
-        task: String,
-    ): HexKey? {
-        if (!isWriteable()) return null
-        val content = Json.encodeToString(WorkflowRunPayload(task = task, workflow = workflowId))
-        val signed = signer.sign(WorkflowTriggerEvent.build(workflowId, content) { workflowChannel(channelId) })
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-        return signed.id
-    }
-
-    /**
-     * Publish a Buzz **workflow definition** (kind-30620) into channel [channelId] on [relay]: an
-     * addressable event whose `d` tag is a freshly-minted workflow UUID (returned here), carrying a
-     * human-readable [name] and the workflow's [yaml] recipe. On a real Buzz relay the relay parses
-     * the YAML and runs it; self-hosted on geode the definition is a named catalog entry the picker
-     * offers and `amy` triggers by id. Returns the new workflow id, or null when the account can't write.
-     */
-    suspend fun publishBuzzWorkflowDef(
-        relay: NormalizedRelayUrl,
-        channelId: String,
-        name: String,
-        yaml: String,
-    ): String? {
-        if (!isWriteable()) return null
-        val workflowId = RandomInstance.randomChars(16)
-        val signed = signer.sign(WorkflowDefEvent.build(workflowId, channelId, yaml, name.ifBlank { null }))
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-        return workflowId
-    }
-
-    /**
-     * Grant a paused Buzz workflow run's approval gate (kind-46030). [runId] is the run id, which
-     * doubles as the approval token (the grant's `d` tag). Resuming lets the runner ship the work.
-     * Publishing to the single group [relay]; the runner discovers the decision by author.
-     */
-    suspend fun approveBuzzWorkflowRun(
-        relay: NormalizedRelayUrl,
-        runId: HexKey,
-        note: String = "",
-    ): HexKey? {
-        if (!isWriteable()) return null
-        val signed = signer.sign(ApprovalGrantEvent.build(runId, note))
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-        return signed.id
-    }
-
-    /** Deny a paused Buzz workflow run's approval gate (kind-46031); the run is terminal (DENIED). */
-    suspend fun denyBuzzWorkflowRun(
-        relay: NormalizedRelayUrl,
-        runId: HexKey,
-        note: String = "",
-    ): HexKey? {
-        if (!isWriteable()) return null
-        val signed = signer.sign(ApprovalDenyEvent.build(runId, note))
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-        return signed.id
-    }
-
-    /**
-     * Upvote a Buzz job [jobId] (authored by [jobAuthor]) — a NIP-25 like (kind-7 `+`) `e`-tagging
-     * the request, `p`-tagging its author and `k`-tagging the reacted kind per NIP-25, and
-     * `h`-scoped to [channelId] so the scheduler (and the board) count it toward priority.
-     */
-    suspend fun upvoteBuzzJob(
-        relay: NormalizedRelayUrl,
-        channelId: String,
-        jobId: HexKey,
-        jobAuthor: HexKey?,
-    ) {
-        if (!isWriteable()) return
-        val template =
-            eventTemplate<ReactionEvent>(ReactionEvent.KIND, ReactionEvent.LIKE) {
-                addUnique(ETag.assemble(jobId, null, null))
-                jobAuthor?.let { addUnique(PTag.assemble(it, null)) }
-                addUnique(arrayOf("k", JobRequestEvent.KIND.toString()))
-                addUnique(GroupIdTag.assemble(channelId))
-            }
-        val signed = signer.sign(template)
-        cache.justConsumeMyOwnEvent(signed)
-        client.publish(signed, setOf(relay))
-    }
-
-    /** Send a kind 9022 leave request to the host relay and drop it from our list. */
-    suspend fun leaveRelayGroup(channel: RelayGroupChannel) {
-        val template = LeaveRequestEvent.build(channel.groupId.id)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-        unfollow(channel)
-    }
-
-    /**
-     * Delete the whole group with a kind 9008 delete-group event (owner/admin only — the relay
-     * enforces this). Unlike [leaveRelayGroup], this destroys the channel for everyone rather than
-     * just removing me; the relay drops the group and its messages. Also drops it from our own list
-     * so it disappears from Messages immediately instead of lingering as a now-dead id.
-     */
-    suspend fun deleteRelayGroup(channel: RelayGroupChannel) {
-        val template = DeleteGroupEvent.build(channel.groupId.id)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-        unfollow(channel)
-        // Remember the deletion so the channel leaves the community's browse list immediately and
-        // stays gone across a restart — the relay drops the group but our cached 39000 metadata (and a
-        // stale re-announced 44100 on a Buzz relay) would otherwise keep it visible.
-        RelayGroupDeletions.markDeleted(channel.groupId)
-    }
-
-    /**
-     * Create a new group on [relay]: kind 9007 (create-group) then kind 9002
-     * (edit-metadata) with the chosen name/visibility, then remember it. Returns
-     * the new group's id.
-     */
-    suspend fun createRelayGroup(
-        relay: NormalizedRelayUrl,
-        groupId: String,
-        name: String,
-        about: String? = null,
-        picture: String? = null,
-        isPrivate: Boolean = false,
-        isClosed: Boolean = false,
-        isHidden: Boolean = false,
-        isRestricted: Boolean = false,
-        hashtags: List<String> = emptyList(),
-        geohashes: List<String> = emptyList(),
-        parent: String? = null,
-        channelType: String? = null,
-    ): GroupId {
-        // The metadata rides the create event as well as the 9002 below. A plain NIP-29 relay takes
-        // its metadata from the 9002 and ignores these tags; Buzz rejects the 9007 outright without
-        // a `name` (see CreateGroupEvent.build), which used to make "create group" on a Buzz relay
-        // publish two events and produce nothing at all.
-        signAndSendPrivatelyOrBroadcast(
-            CreateGroupEvent.build(
-                groupId = groupId,
-                name = name,
-                about = about,
-                visibility = if (isPrivate) BUZZ_VISIBILITY_PRIVATE else BUZZ_VISIBILITY_OPEN,
-                channelType = channelType,
-            ),
-        ) { listOf(relay) }
-
-        val edit =
-            EditMetadataEvent.build(
-                groupId,
-                name = name,
-                about = about,
-                picture = picture,
-                status = relayGroupStatus(isPrivate, isClosed, isHidden, isRestricted),
-                hashtags = hashtags,
-                geohashes = geohashes,
-                parent = parent,
-            )
-        signAndSendPrivatelyOrBroadcast(edit) { listOf(relay) }
-
-        val id = GroupId(groupId, relay)
-        follow(LocalCache.getOrCreateRelayGroupChannel(id))
-        return id
-    }
-
-    /**
-     * The set of NIP-29 status flags to emit on a kind-9002 metadata event. Flags are
-     * presence-only — public/open/visible/unrestricted are simply the ABSENCE of their
-     * restrictive counterpart — so only the enabled restrictive flags are added.
-     */
-    private fun relayGroupStatus(
-        isPrivate: Boolean,
-        isClosed: Boolean,
-        isHidden: Boolean,
-        isRestricted: Boolean,
-    ): Set<GroupMetadataEvent.GroupStatus> =
-        buildSet {
-            if (isPrivate) add(GroupMetadataEvent.GroupStatus.PRIVATE)
-            if (isClosed) add(GroupMetadataEvent.GroupStatus.CLOSED)
-            if (isHidden) add(GroupMetadataEvent.GroupStatus.HIDDEN)
-            if (isRestricted) add(GroupMetadataEvent.GroupStatus.RESTRICTED)
-        }
-
-    /** Post a kind 11 thread (forum-style) to the group, scoped by its `h` tag. */
-    suspend fun postRelayGroupThread(
-        channel: RelayGroupChannel,
-        title: String,
-        body: String,
-    ) {
-        val template =
-            ThreadEvent.build(body, title) {
-                hTag(channel.groupId.id)
-                previous(channel.previousEventRefs(pubKey))
-            }
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /** Mint a kind 9009 invite code for the group (admin/moderator only). */
-    suspend fun createRelayGroupInvite(
-        channel: RelayGroupChannel,
-        code: String,
-    ) {
-        val template = CreateInviteEvent.build(channel.groupId.id, code)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /**
-     * Replace the group's pinned-message list with a kind 9010 update-pin-list event
-     * (admin/moderator only). NIP-29 carries the FULL list, so the relay applies it and
-     * republishes the kind-39005 [com.vitorpamplona.quartz.nip29RelayGroups.metadata.GroupPinnedEvent].
-     */
-    suspend fun updateRelayGroupPins(
-        channel: RelayGroupChannel,
-        pinnedEventIds: List<HexKey>,
-    ) {
-        val template = UpdatePinListEvent.build(channel.groupId.id, pinnedEventIds)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /** Pin [eventId] by appending it to the current list (no-op if already pinned). */
-    suspend fun pinRelayGroupMessage(
-        channel: RelayGroupChannel,
-        eventId: HexKey,
-    ) {
-        if (channel.isPinned(eventId)) return
-        updateRelayGroupPins(channel, channel.pinnedEventIds + eventId)
-    }
-
-    /** Unpin [eventId] by removing it from the current list (no-op if not pinned). */
-    suspend fun unpinRelayGroupMessage(
-        channel: RelayGroupChannel,
-        eventId: HexKey,
-    ) {
-        if (!channel.isPinned(eventId)) return
-        updateRelayGroupPins(channel, channel.pinnedEventIds - eventId)
-    }
-
-    /** Kick [pubkey] out of the group with a kind 9001 remove-user event (moderator only). */
-    suspend fun removeRelayGroupUser(
-        channel: RelayGroupChannel,
-        pubkey: HexKey,
-    ) {
-        val template = RemoveUserEvent.build(channel.groupId.id, listOf(pubkey))
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /**
-     * Add [pubkey] to the group (or change its roles) with a kind 9000 put-user
-     * event (moderator only). Pass an empty [roles] list for a plain member.
-     */
-    suspend fun putRelayGroupUser(
-        channel: RelayGroupChannel,
-        pubkey: HexKey,
-        roles: List<String>,
-    ) {
-        // Buzz ignores the roles inside the `p` tag and reads a top-level `role` tag instead, in its
-        // own vocabulary — so map ours onto its set before sending. Anything it cannot parse fails
-        // the whole put-user, which is why an unmapped role must become `member` rather than travel.
-        val buzzRole =
-            if (BuzzRelayDialect.isBuzz(channel.groupId.relayUrl)) {
-                when {
-                    roles.any { it.equals(RelayGroupMembership.ROLE_ADMIN, true) } -> BUZZ_ROLE_ADMIN
-                    else -> BUZZ_ROLE_MEMBER
-                }
-            } else {
-                null
-            }
-        val template = PutUserEvent.build(channel.groupId.id, listOf(pubkey to roles), buzzRole = buzzRole)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /**
-     * Add [pubkey] to a Buzz **community** (the whole relay/tenant, not one channel) via the
-     * relay-admin add-member command (kind 9030). Owner/admin only — the relay validates the
-     * sender's role and, on a new insert, updates its NIP-43 membership list (13534). Published to
-     * [relay] with no channel scope.
-     */
-    suspend fun addCommunityMember(
-        relay: NormalizedRelayUrl,
-        pubkey: HexKey,
-        role: String? = null,
-    ) {
-        signAndSendPrivatelyOrBroadcast(RelayAdminAddMemberEvent.build(pubkey, role)) { listOf(relay) }
-    }
-
-    /** Remove [pubkey] from a Buzz community via the relay-admin remove-member command (kind 9031). */
-    suspend fun removeCommunityMember(
-        relay: NormalizedRelayUrl,
-        pubkey: HexKey,
-    ) {
-        signAndSendPrivatelyOrBroadcast(RelayAdminRemoveMemberEvent.build(pubkey)) { listOf(relay) }
-    }
-
-    /**
-     * Edit the group's relay-signed metadata with a kind 9002 event (admin only).
-     *
-     * NIP-29 §Subgroups makes the metadata edit a full replacement of the hierarchy
-     * links: a 9002 with no `parent` tag re-roots the group, and one that drops any
-     * existing `child` is rejected by the relay. So unless the caller is explicitly
-     * re-parenting, we re-carry the group's current [parent] and full [children] list
-     * from its latest known metadata to keep the tree intact across a plain name/flag
-     * edit. Pass an explicit value to change them.
-     */
-    suspend fun editRelayGroupMetadata(
-        channel: RelayGroupChannel,
-        name: String?,
-        about: String?,
-        picture: String?,
-        isPrivate: Boolean,
-        isClosed: Boolean,
-        isHidden: Boolean,
-        isRestricted: Boolean,
-        hashtags: List<String> = emptyList(),
-        geohashes: List<String> = emptyList(),
-        parent: String? = channel.parentGroupId(),
-        children: List<String> = channel.childGroupIds(),
-    ) {
-        // On a Buzz relay, visibility rides a `visibility` ("open"/"private") tag — the relay does NOT
-        // read NIP-29's `private` status flag — so a Buzz channel's visibility only actually changes on
-        // edit when we send that tag. A plain NIP-29 relay ignores it and honours the status flag.
-        val isBuzz = BuzzRelayDialect.isBuzz(channel.groupId.relayUrl)
-        val template =
-            EditMetadataEvent.build(
-                channel.groupId.id,
-                name = name,
-                about = about,
-                picture = picture,
-                status = relayGroupStatus(isPrivate, isClosed, isHidden, isRestricted),
-                hashtags = hashtags,
-                geohashes = geohashes,
-                parent = parent,
-                children = children,
-                visibility = if (isBuzz) (if (isPrivate) BUZZ_VISIBILITY_PRIVATE else BUZZ_VISIBILITY_OPEN) else null,
-            )
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
-    }
-
-    /**
-     * Archive or unarchive a Buzz channel (a minimal kind-9002 carrying only the `archived` tag). The
-     * relay hides an archived channel from the sidebar and stamps the 39000, but keeps it and its
-     * history — the reversible counterpart to [deleteRelayGroup]. Admin/owner only; the relay enforces.
-     */
-    suspend fun archiveRelayGroup(
-        channel: RelayGroupChannel,
-        archived: Boolean,
-    ) {
-        val template = EditMetadataEvent.build(channel.groupId.id, archived = archived)
-        signAndSendPrivatelyOrBroadcast(template) { channel.relays().toList() }
     }
 
     suspend fun follow(community: AddressableNote) = sendMyPublicAndPrivateOutbox(communityList.follow(community))
