+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/buzz/BuzzNewDmViewModel.kt
@@ -194,7 +194,7 @@ class BuzzNewDmViewModel : ViewModel() {
         _status.value = Status.Sending
         viewModelScope.launch(Dispatchers.IO) {
             try {
-                val channelId = account.openBuzzDm(relay, others)
+                val channelId = account.relayGroups.openBuzzDm(relay, others)
                 val groupId = channelId?.let { GroupId(it, relay) }
                 withContext(Dispatchers.Main) { onOpened(groupId) }
             } catch (e: CancellationException) {
